--- 6/SBomberProject/SBomberProject/Bomb.h
@@ -1,16 +1,0 @@
-#pragma once
-
-#include "DynamicObject.h"
-
-class Bomb : public DynamicObject
-{
-public:
-
-	static const uint16_t BombCost = 10; // стоимость бомбы в очках
-
-	void Draw() override;
-
-private:
-
-};
-
