--- 6/SBomberProject/SBomberProject/Ground.h
@@ -1,46 +1,0 @@
-#pragma once
-
-#include <vector>
-
-#include "GameObject.h"
-
-//=============================================================================================================
-
-enum CraterSize 
-{ 
-	SMALL_CRATER_SIZE = 9 
-};
-
-//=============================================================================================================
-
-class Crater : public GameObject
-{
-public:
-
-	bool __fastcall isInside(double xn) const;
-	
-	void Draw() override;
-
-private:
-
-};
-
-//=============================================================================================================
-
-class Ground : public GameObject
-{
-public:
-
-	Ground() { }
-
-	void Draw() override;
-
-	void __fastcall AddCrater(double xn);
-
-private:
-
-	bool __fastcall isInsideAnyCrater(double x) const;
-
-	std::vector<Crater> vecCrates;
-};
-
