--- 6/SBomberProject/SBomberProject/House.cpp
@@ -1,48 +1,0 @@
-
-#include <iostream>
-
-#include "House.h"
-#include "MyTools.h"
-
-using namespace std;
-using namespace MyTools;
-
-bool House::isInside(double x1, double x2) const
-{
-	const double XBeg = x + 2;
-	const double XEnd = x + width - 1;
-
-	if (x1 < XBeg && x2 > XEnd)
-	{
-		return true;
-	}
-
-	if (x1 > XBeg && x1 < XEnd)
-	{
-		return true;
-	}
-
-	if (x2 > XBeg && x2 < XEnd)
-	{
-		return true;
-	}
-
-	return false;
-}
-
-void House::Draw() 
-{
-	MyTools::ScreenSingleton::getInstance().SetColor(CC_Yellow);
-	MyTools::ScreenSingleton::getInstance().GotoXY(x, y - 5);
-	cout << "  ########  ";
-	MyTools::ScreenSingleton::getInstance().GotoXY(x, y - 4);
-	cout << "##        ##";
-	MyTools::ScreenSingleton::getInstance().GotoXY(x, y - 3);
-	cout << "############";
-	MyTools::ScreenSingleton::getInstance().GotoXY(x, y - 2);
-	cout << "#          #";
-	MyTools::ScreenSingleton::getInstance().GotoXY(x, y - 1);
-	cout << "#          #";
-	MyTools::ScreenSingleton::getInstance().GotoXY(x, y);
-	cout << "############";
-}
