--- 6/SBomberProject/SBomberProject/SBomberProject.cpp
@@ -1,38 +1,0 @@
-﻿
-#include <conio.h>
-
-#include "SBomber.h"
-#include "MyTools.h"
-
-using namespace std;
-
-//========================================================================================================================
-
-int main(void)
-{
-    MyTools::FileLoggerSingleton::getInstance().OpenLogFile("log.txt");
-
-    SBomber game;
-
-    do {
-        game.TimeStart();
-
-        if (_kbhit())
-        {
-            game.ProcessKBHit();
-        }
-
-        MyTools::ScreenSingleton::getInstance().ClrScr();
-
-        game.DrawFrame();
-        game.MoveObjects();
-        game.CheckObjects();
-
-        game.TimeFinish();
-
-    } while (!game.GetExitFlag());
-
-    MyTools::FileLoggerSingleton::getInstance().CloseLogFile();
-
-    return 0;
-}
