--- 6/SBomberProject/SBomberProject/x64/Debug/SBomberProject.Build.CppClean.log
@@ -1,20 +1,0 @@
-c:\users\danil\desktop\007_patterns\1\sbomberproject\sbomberproject\x64\debug\vc143.pdb
-c:\users\danil\desktop\007_patterns\1\sbomberproject\sbomberproject\x64\debug\vc143.idb
-c:\users\danil\desktop\007_patterns\1\sbomberproject\sbomberproject\x64\debug\tank.obj
-c:\users\danil\desktop\007_patterns\1\sbomberproject\sbomberproject\x64\debug\sbomberproject.obj
-c:\users\danil\desktop\007_patterns\1\sbomberproject\sbomberproject\x64\debug\sbomber.obj
-c:\users\danil\desktop\007_patterns\1\sbomberproject\sbomberproject\x64\debug\plane.obj
-c:\users\danil\desktop\007_patterns\1\sbomberproject\sbomberproject\x64\debug\mytools.obj
-c:\users\danil\desktop\007_patterns\1\sbomberproject\sbomberproject\x64\debug\levelgui.obj
-c:\users\danil\desktop\007_patterns\1\sbomberproject\sbomberproject\x64\debug\house.obj
-c:\users\danil\desktop\007_patterns\1\sbomberproject\sbomberproject\x64\debug\ground.obj
-c:\users\danil\desktop\007_patterns\1\sbomberproject\sbomberproject\x64\debug\bomb.obj
-c:\users\danil\desktop\007_patterns\1\sbomberproject\sbomberproject\x64\debug\sbomberproject.ilk
-c:\users\danil\desktop\007_patterns\1\sbomberproject\x64\debug\sbomberproject.exe
-c:\users\danil\desktop\007_patterns\1\sbomberproject\x64\debug\sbomberproject.pdb
-c:\users\danil\desktop\007_patterns\1\sbomberproject\sbomberproject\x64\debug\sbomberproject.tlog\cl.command.1.tlog
-c:\users\danil\desktop\007_patterns\1\sbomberproject\sbomberproject\x64\debug\sbomberproject.tlog\cl.read.1.tlog
-c:\users\danil\desktop\007_patterns\1\sbomberproject\sbomberproject\x64\debug\sbomberproject.tlog\cl.write.1.tlog
-c:\users\danil\desktop\007_patterns\1\sbomberproject\sbomberproject\x64\debug\sbomberproject.tlog\link.command.1.tlog
-c:\users\danil\desktop\007_patterns\1\sbomberproject\sbomberproject\x64\debug\sbomberproject.tlog\link.read.1.tlog
-c:\users\danil\desktop\007_patterns\1\sbomberproject\sbomberproject\x64\debug\sbomberproject.tlog\link.write.1.tlog
