--- 6/SBomberProject/SBomberProject/x64/Debug/SBomberProject.exe.recipe
@@ -1,11 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project>
-  <ProjectOutputs>
-    <ProjectOutput>
-      <FullPath>C:\Users\Danil\Desktop\007_Patterns\6\SBomberProject\x64\Debug\SBomberProject.exe</FullPath>
-    </ProjectOutput>
-  </ProjectOutputs>
-  <ContentFiles />
-  <SatelliteDlls />
-  <NonRecipeFileRefs />
-</Project>
