+++ 7/SBomberProject/Debug/log.txt
