+++ 7/SBomberProject/SBomberProject/Bomb.cpp
@@ -1,0 +1,15 @@
+
+#include <iostream>
+
+#include "Bomb.h"
+#include "MyTools.h"
+
+using namespace std;
+using namespace MyTools;
+
+void Bomb::Draw() const
+{
+    MyTools::SetColor(CC_LightMagenta);
+    GotoXY(x, y);
+    cout << "*";
+}
