+++ 7/SBomberProject/SBomberProject/Bomb.h
@@ -1,0 +1,16 @@
+#pragma once
+
+#include "DynamicObject.h"
+
+class Bomb : public DynamicObject
+{
+public:
+
+	static const uint16_t BombCost = 10; // стоимость бомбы в очках
+
+	void Draw() const override;
+
+private:
+
+};
+
