+++ 7/SBomberProject/SBomberProject/DynamicObject.h
