+++ 7/SBomberProject/SBomberProject/Plane.h
@@ -1,0 +1,15 @@
+#pragma once
+
+#include "DynamicObject.h"
+
+class Plane : public DynamicObject {
+public:
+
+    void Draw() const override;
+
+    inline void ChangePlaneY(double dy) { yDirection += dy; }
+
+private:
+
+};
+
