+++ 7/SBomberProject/SBomberProject/SBomber.h
@@ -1,0 +1,33 @@
+#pragma once
+
+#include <vector>
+
+#include "LevelGUI.h"
+#include "Plane.h"
+#include "Bomb.h"
+#include "Ground.h"
+#include "Tank.h"
+
+
+class SBomber
+{
+public:
+
+    SBomber();
+    ~SBomber();
+    
+    bool GetExitFlag() const;
+
+    void ProcessKBHit();
+    void TimeStart();
+    void TimeFinish();
+
+    void DrawFrame();
+    void MoveObjects();
+    void CheckObjects();
+
+private:
+
+    class SBomberImpl;
+    SBomberImpl* pImpl;
+};
