+++ 7/SBomberProject/SBomberProject/SBomberProject.cpp
@@ -1,0 +1,37 @@
+﻿
+#include <conio.h>
+
+#include "SBomber.h"
+#include "MyTools.h"
+
+using namespace std;
+
+//========================================================================================================================
+
+MyTools::FileLogger logger("log.txt"); //a definition of the class "FileLogger" is in the files "MyTools.h" and "MyTools.cpp"
+
+int main(void)
+{
+
+    SBomber game;
+
+    do {
+        game.TimeStart();
+
+        if (_kbhit())
+        {
+            game.ProcessKBHit();
+        }
+
+        MyTools::ClrScr();
+
+        game.DrawFrame();
+        game.MoveObjects();
+        game.CheckObjects();
+
+        game.TimeFinish();
+
+    } while (!game.GetExitFlag());
+
+    return 0;
+}
