+++ 7/SBomberProject/SBomberProject/Tank.h
@@ -1,0 +1,21 @@
+#pragma once
+
+#include <stdint.h>
+
+#include "DestroyableGroundObject.h"
+
+class Tank : public DestroyableGroundObject
+{
+public:
+
+	bool __fastcall isInside(double x1, double x2) const override;
+
+	inline uint16_t GetScore() const override { return score; }
+
+	void Draw() const override;
+
+private:
+
+	const uint16_t score = 30;
+};
+
