+++ 7/SBomberProject/SBomberProject/x64/Debug/SBomberProject.log
@@ -1,0 +1,4 @@
+﻿  SBomber.cpp
+  SBomberProject.cpp
+  Generating Code...
+  SBomberProject.vcxproj -> C:\Users\Danil\Desktop\007_Patterns\7\SBomberProject\x64\Debug\SBomberProject.exe
